fix role 1

+++ backend/routes/api.php
@@ -43,20 +43,18 @@ Route::group([
 
 });
 
-Route::middleware(['auth', 'XSSProtection', 'role:Admin'])->group(function () {
-    // Route::middleware(['auth', 'XSSProtection', 'role:QLHT'])->group(function () {
-        Route::get('/student', [StudentController::class, 'index']);
-        Route::get('student/{id}', [StudentController::class, 'show']);
-        Route::post('student', [StudentController::class, 'store']);
-        Route::put('student/{id}', [StudentController::class, 'update']);
-        Route::delete('student/{id}', [StudentController::class, 'destroy']);
+Route::middleware(['auth', 'XSSProtection', 'role:Admin|QLHT'])->group(function () {
+    Route::get('/student', [StudentController::class, 'index']);
+    Route::get('student/{id}', [StudentController::class, 'show']);
+    Route::post('student', [StudentController::class, 'store'])->middleware('permission:add student');
+    Route::put('student/{id}', [StudentController::class, 'update'])->middleware('permission:edit student');
+    Route::delete('student/{id}', [StudentController::class, 'destroy'])->middleware('permission:delete student');
 
-        Route::get('/user', [UserController::class, 'index']);
-        Route::get('user/{id}', [UserController::class, 'show']);
-        Route::post('user', [UserController::class, 'store']);
-        Route::put('user/{id}', [UserController::class, 'update']);
-        Route::delete('user/{id}', [UserController::class, 'destroy']);
-    // });
+    Route::get('/user', [UserController::class, 'index']);
+    Route::get('user/{id}', [UserController::class, 'show']);
+    Route::post('user', [UserController::class, 'store'])->middleware(['permission:add user', 'isAdmin']);
+    Route::put('user/{id}', [UserController::class, 'update'])->middleware('permission:edit user');
+    Route::delete('user/{id}', [UserController::class, 'destroy'])->middleware('permission:delete user');
 });
